refactor: restructure 'argocd-util' commands
