+++ docs/operator-manual/server-commands/argocd-util_app.md
@@ -1,0 +1,21 @@
+## argocd-util app
+
+Manage applications configuration
+
+```
+argocd-util app [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for app
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+* [argocd-util app diff-reconcile-results](argocd-util_app_diff-reconcile-results.md)	 - Compare results of two reconciliations and print diff.
+* [argocd-util app generate-spec](argocd-util_app_generate-spec.md)	 - Generate declarative config for an application
+* [argocd-util app get-reconcile-results](argocd-util_app_get-reconcile-results.md)	 - Reconcile all applications and stores reconciliation summary in the specified file.
+
