+++ docs/operator-manual/server-commands/argocd-util_app_diff-reconcile-results.md
@@ -1,0 +1,18 @@
+## argocd-util app diff-reconcile-results
+
+Compare results of two reconciliations and print diff.
+
+```
+argocd-util app diff-reconcile-results PATH1 PATH2 [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for diff-reconcile-results
+```
+
+### SEE ALSO
+
+* [argocd-util app](argocd-util_app.md)	 - Manage applications configuration
+
