+++ docs/operator-manual/server-commands/argocd-util_cluster.md
@@ -1,0 +1,20 @@
+## argocd-util cluster
+
+Manage clusters configuration
+
+```
+argocd-util cluster [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for cluster
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+* [argocd-util cluster generate-spec](argocd-util_cluster_generate-spec.md)	 - Generate declarative config for a cluster
+* [argocd-util cluster kubeconfig](argocd-util_cluster_kubeconfig.md)	 - Generates kubeconfig for the specified cluster
+
