+++ docs/operator-manual/server-commands/argocd-util_cluster_kubeconfig.md
@@ -1,16 +1,14 @@
-## argocd-util settings
+## argocd-util cluster kubeconfig
 
-Provides set of commands for settings validation and troubleshooting
+Generates kubeconfig for the specified cluster
 
 ```
-argocd-util settings [flags]
+argocd-util cluster kubeconfig CLUSTER_URL OUTPUT_PATH [flags]
 ```
 
 ### Options
 
 ```
-      --argocd-cm-path string          Path to local argocd-cm.yaml file
-      --argocd-secret-path string      Path to local argocd-secret.yaml file
       --as string                      Username to impersonate for the operation
       --as-group stringArray           Group to impersonate for the operation, this flag can be repeated to specify multiple groups.
       --certificate-authority string   Path to a cert file for the certificate authority
@@ -18,10 +16,9 @@ argocd-util settings [flags]
       --client-key string              Path to a client key file for TLS
       --cluster string                 The name of the kubeconfig cluster to use
       --context string                 The name of the kubeconfig context to use
-  -h, --help                           help for settings
+  -h, --help                           help for kubeconfig
       --insecure-skip-tls-verify       If true, the server's certificate will not be checked for validity. This will make your HTTPS connections insecure
       --kubeconfig string              Path to a kube config. Only required if out-of-cluster
-      --load-cluster-settings          Indicates that config map and secret should be loaded from cluster unless local file path is provided
   -n, --namespace string               If present, the namespace scope for this CLI request
       --password string                Password for basic authentication to the API server
       --request-timeout string         The length of time to wait before giving up on a single server request. Non-zero values should contain a corresponding time unit (e.g. 1s, 2m, 3h). A value of zero means don't timeout requests. (default "0")
@@ -34,8 +31,5 @@ argocd-util settings [flags]
 
 ### SEE ALSO
 
-* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
-* [argocd-util settings rbac](argocd-util_settings_rbac.md)	 - Validate and test RBAC configuration
-* [argocd-util settings resource-overrides](argocd-util_settings_resource-overrides.md)	 - Troubleshoot resource overrides
-* [argocd-util settings validate](argocd-util_settings_validate.md)	 - Validate settings
+* [argocd-util cluster](argocd-util_cluster.md)	 - Manage clusters configuration
