+++ docs/operator-manual/server-commands/argocd-util_proj.md
@@ -1,0 +1,21 @@
+## argocd-util proj
+
+Manage projects configuration
+
+```
+argocd-util proj [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for proj
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+* [argocd-util proj generate-allow-list](argocd-util_proj_generate-allow-list.md)	 - Generates project allow list from the specified clusterRole file
+* [argocd-util proj generate-spec](argocd-util_proj_generate-spec.md)	 - Generate declarative config for a project
+* [argocd-util proj update-role-policy](argocd-util_proj_update-role-policy.md)	 - Implement bulk project role update. Useful to back-fill existing project policies or remove obsolete actions.
+
