+++ docs/operator-manual/server-commands/argocd-util_repo.md
@@ -1,0 +1,19 @@
+## argocd-util repo
+
+Manage repositories configuration
+
+```
+argocd-util repo [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for repo
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+* [argocd-util repo generate-spec](argocd-util_repo_generate-spec.md)	 - Generate declarative config for a repo
+
