+++ docs/operator-manual/server-commands/argocd-util_settings_rbac_can.md
@@ -1,12 +1,53 @@
-## argocd-util settings
+## argocd-util settings rbac can
 
-Provides set of commands for settings validation and troubleshooting
+Check RBAC permissions for a role or subject
+
+### Synopsis
+
+
+Check whether a given role or subject has appropriate RBAC permissions to do
+something.
+
 
 ```
-argocd-util settings [flags]
+argocd-util settings rbac can ROLE/SUBJECT ACTION RESOURCE [SUB-RESOURCE] [flags]
+```
+
+### Examples
+
+```
+
+# Check whether role some:role has permissions to create an application in the
+# 'default' project, using a local policy.csv file
+argocd-util rbac can some:role create application 'default/app' --policy-file policy.csv
+
+# Policy file can also be K8s config map with data keys like argocd-rbac-cm,
+# i.e. 'policy.csv' and (optionally) 'policy.default'
+argocd-util rbac can some:role create application 'default/app' --policy-file argocd-rbac-cm.yaml
+
+# If --policy-file is not given, the ConfigMap 'argocd-rbac-cm' from K8s is
+# used. You need to specify the argocd namespace, and make sure that your
+# current Kubernetes context is pointing to the cluster Argo CD is running in
+argocd-util rbac can some:role create application 'default/app' --namespace argocd
+
+# You can override a possibly configured default role
+argocd-util rbac can someuser create application 'default/app' --default-role role:readonly
+
+
 ```
 
 ### Options
+
+```
+      --default-role string   name of the default role to use
+  -h, --help                  help for can
+      --policy-file string    path to the policy file to use
+  -q, --quiet                 quiet mode - do not print results to stdout
+      --strict                whether to perform strict check on action and resource names (default true)
+      --use-builtin-policy    whether to also use builtin-policy (default true)
+```
+
+### Options inherited from parent commands
 
 ```
       --argocd-cm-path string          Path to local argocd-cm.yaml file
@@ -18,7 +59,6 @@ argocd-util settings [flags]
       --client-key string              Path to a client key file for TLS
       --cluster string                 The name of the kubeconfig cluster to use
       --context string                 The name of the kubeconfig context to use
-  -h, --help                           help for settings
       --insecure-skip-tls-verify       If true, the server's certificate will not be checked for validity. This will make your HTTPS connections insecure
       --kubeconfig string              Path to a kube config. Only required if out-of-cluster
       --load-cluster-settings          Indicates that config map and secret should be loaded from cluster unless local file path is provided
@@ -34,8 +74,5 @@ argocd-util settings [flags]
 
 ### SEE ALSO
 
-* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
 * [argocd-util settings rbac](argocd-util_settings_rbac.md)	 - Validate and test RBAC configuration
-* [argocd-util settings resource-overrides](argocd-util_settings_resource-overrides.md)	 - Troubleshoot resource overrides
-* [argocd-util settings validate](argocd-util_settings_validate.md)	 - Validate settings
